% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_template_format_db.R
\name{build_format_db}
\alias{build_format_db}
\alias{build_template_format_db}
\title{Build template definition format}
\usage{
build_template_format_db(format_name, format_version = "0.1.0")
}
\arguments{
\item{format_name}{Name of the format.}

\item{format_version}{Version of the format (by default 0.1.0). To know more about the definition format take a look the corresponding specification section.}
}
\value{
A list of 9 elements
}
\description{
Build template definition format.
}
\section{Specifications}{

\describe{
\item{Format version}
{The version format is always defined with 3 numbers each separated by a dot: <major>.<minor>.<patch>.
The first number (<major>) is related to a major updated. It's when you have added many new features or conceptual changes impacted whose directly to the user interface (typically the new user interface is not compatible with the previous).
The second number (<minor>)  is when you add functionality in a backwards-compatible manner.
The third number (<patch>) is related to bug resolutions and more preciously when you make backwards-compatible bug fixes.}
}
}

\examples{
# Built a template of the definition format with a version 0.1.0 (default setting)
build_format_db <- ("name_of_my_format")
# Built a template of the definition format with a version 1.0.1
build_format_db <- ("name_of_my_format", format_version = "1.0.1")
}
\references{
\url{TO DO}
}
\author{
Mathieu Depetris, \email{mathieu.depetris@ird.fr}
}
