% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checking_data.R
\name{checking_data}
\alias{checking_data}
\title{Validation process for data}
\usage{
checking_data(obj, format_db, ignore_case_in_codelist = TRUE,
  report = "files", report_dir = tempdir(), text_file_sep = ";",
  text_file_dec = ".", file_name_slot)
}
\arguments{
\item{obj}{Path of the file or R's object that contain data.}

\item{format_db}{Path of the file or R's object that contain the definition format.}

\item{ignore_case_in_codelist}{Ignored cases (upper or lower) in the codelist. By default yes (TRUE).}

\item{report}{Selected format for the report. You could choose between files, list or both. By default files selected.}

\item{report_dir}{Location of export directory for the report. By default, the function uses the temporary directory.}

\item{text_file_sep}{If the argument obj is a csv file, specify here the field separator of it. By default the separator is ";".}

\item{text_file_dec}{If the argument obj is a csv file, specify here the string use for decimal points. By default the decimal is ".".}

\item{file_name_slot}{If the argument obj is a csv file, specify here the name of the slot. By default not provided.}
}
\value{
The function returns file(s) location or list(s) name(s) of the report element(s).
}
\description{
Validation process for data.
}
\references{
\url{https://github.com/OB7-IRD/dqassess/blob/master/R/checking_data.R}
}
\author{
Mathieu Depetris, \email{mathieu.depetris@ird.fr}
}
