% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dqassess.R
\docType{package}
\name{dqassess}
\alias{dqassess}
\alias{dqassess-package}
\title{Data quality assessment}
\description{
This package contain serveral functions in order to evaluate the conformity of data against a data format description.
It attempt to help on data error detection concerning for example the data structures, formats or code lists.
This R package was developped within the work package 5 of the FRAMEWORK CONTRACT - MARE/2016/22 "Strengthening REgional COoperation in the area of LArge PElagic fisheries data collection" (RECOLAPE project).
}
\section{List of available functions}{

\itemize{
\item{\code{\link{build_template_format_db}}}
\item{\code{\link{read_format_db}}}
\item{\code{\link{write_format_db_excel}}}
\item{\code{\link{checking_data}}}
}
}

\section{Vignettes link}{

Several vignettes was made for explain the methodology behind the function and for give examples of application:
\itemize{
\item{\code{\link{../vignettes/dqassess_definition_data_format.html}}}
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/OB7-IRD/dqassess}
  \item Report bugs at \url{https://github.com/OB7-IRD/dqassess/issues}
}

}
\author{
Mathieu Depetris, \email{mathieu.depetris@ird.fr}

Norbert Billet, \email{norbert.billet@ifremer.fr}
}
