% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_format_db.R
\name{read_format_db}
\alias{read_format_db}
\title{Read a format file definition}
\usage{
read_format_db(input_file_path)
}
\arguments{
\item{input_file_path}{Path of the Excel file.}
}
\value{
A list of elements from the format file definition.
}
\description{
Read a format file definition from an Excel file
}
\examples{
read_format_db <- ("path_of_the_file\\\\file_name.xlsx")
}
\references{
\url{https://github.com/OB7-IRD/dqassess/blob/master/R/read_format_db.R}
}
\author{
Mathieu Depetris, \email{mathieu.depetris@ird.fr}
}
