% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_format_db_excel.R
\name{write_format_db_excel}
\alias{write_format_db_excel}
\title{Generation excel definition format}
\usage{
write_format_db_excel(format_db, output_file_path_dir = tempdir(),
  output_file_name = paste("formatdef", format(Sys.time(),
  "\%Y\%m\%d_\%H\%M\%S"), sep = "_"), output_file_ext = "xlsx")
}
\arguments{
\item{format_db}{Name of the R's element where is stock the format definition.}

\item{output_file_path_dir}{Location of export directory for the format definition. By default, the function uses the temporary directory.}

\item{output_file_name}{Name of the file exported. By default, the function use this syntax: formatdef_\%Y\%m\%d_\%H\%M\%S (check formats of \code{\link[base]{strptime}} function in R).}

\item{output_file_ext}{Extension of the file exported. You have to choose between xlsx for Excel 2007 OOXML format, or xls for Excel 95 binary format. Be default file are exported in xlsx.}
}
\value{
The function return the path of the exported file.
}
\description{
Generation of a definition format in a excel file.
}
\references{
\url{TO DO}
}
\author{
Mathieu Depetris, \email{mathieu.depetris@ird.fr}
}
